% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-data-recruitment.R
\name{bbd_chk_data_recruitment}
\alias{bbd_chk_data_recruitment}
\title{Check recruitment data structure}
\format{
The data must follow the requirements:
\describe{
\item{PopulationName}{Name of the herd or population}
\item{Year}{The calendar year the observation occurred. Must be a positive
integer.}
\item{Month}{The calendar month the observation occurred. Must be an integer
between 1 and 12.}
\item{Day}{The day the observation occurred. Must be an integer between 1 and
31.}
\item{Cows}{The total number of cows counted in each group in a survey/year.
Must be a positive integer.}
\item{Bulls}{The total number of bulls counted in each group in a
survey/year. Must be a positive integer}
\item{UnknownAdults}{The total number of adults counted that the sex could
not be identified in each group in a survey/year. Must be a positive
integer.}
\item{Yearlings}{The total number of yearlings that did not have the sex
identified in each group in a survey/year. Must be a positive integer. }
\item{Calves}{The total number of calves counted in each group in a
survey/year. Must be a positive integer.}
}
}
\usage{
bbd_chk_data_recruitment(
  data,
  x_name = deparse(substitute(data)),
  multi_population = FALSE,
  allow_missing = FALSE
)
}
\arguments{
\item{data}{The data.frame to check.}

\item{x_name}{A string of the name of the data.frame.}

\item{multi_population}{A flag indicating whether to accept multiple populations.}

\item{allow_missing}{A flag indicating whether to accept missing values for 'Cows', 'Bulls', 'UnknownAdults', 'Yearlings', and 'Calves'.}
}
\value{
An invisible copy of the original data.frame.
}
\description{
The data must follow all requirements to not error. This format is required
for usage of the bbou suite of tools.
}
\examples{
bbd_chk_data_recruitment(bbourecruit_a)
bbd_chk_data_recruitment(bbourecruit_b)
bbd_chk_data_recruitment(bbourecruit_c)
# this example will error as it doesn't follow the requirements
x <- bbourecruit_a
x[1, 4] <- 32L
try(bbd_chk_data_recruitment(x))
}
