% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-data-survival.R
\name{bbd_chk_data_survival}
\alias{bbd_chk_data_survival}
\title{Check survival data structure}
\usage{
bbd_chk_data_survival(data, x_name = deparse(substitute(data)))
}
\arguments{
\item{data}{A data frame.}

\item{x_name}{Name of data frame.}
}
\value{
An invisible data frame.
}
\description{
Check survival data structure
}
\examples{
bbd_chk_data_survival(bbousurv_a)
}
