% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbourecruit.R
\docType{data}
\name{bbourecruit_a}
\alias{bbourecruit_a}
\title{Sample Data for Population A}
\format{
A tibble
\describe{
\item{PopulationName}{Name of the population}
\item{Year}{The year}
\item{Month}{The month}
\item{Day}{The day}
\item{Cows}{The number of cows}
\item{Bulls}{The number of bulls}
\item{UnknownAdults}{The number of adults which could not be sexed}
\item{Yearlings}{The number of yearlings}
\item{Calves}{The number of calves}
}
}
\usage{
bbourecruit_a
}
\description{
The data contains recruitment information for a boreal caribou population B
which spans 26 years.
}
\details{
The data is released under the \href{https://open.alberta.ca/licence}{Open Government Licence - Alberta}
}
\keyword{datasets}
