% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbourecruit.R
\docType{data}
\name{bbourecruit_b}
\alias{bbourecruit_b}
\title{Sample Data for Population B}
\format{
A tibble with columns:
\describe{
\item{PopulationName}{Name of the population}
\item{Year}{The year the observation occurred}
\item{Month}{The month the observation occurred}
\item{Day}{The day the observation occurred}
\item{Cows}{The number of cows counted in the group}
\item{Bulls}{The number of bulls counted in the group}
\item{UnknownAdults}{The number of adults which could not be sexed in the
group}
\item{Yearlings}{The number of yearlings counted in the group}
\item{Calves}{The number of calves counted in the group}
}
}
\usage{
bbourecruit_b
}
\description{
The data contains recruitment information for boreal caribou population B
which spans 14 years.
}
\details{
The data is released under the \href{https://open.alberta.ca/licence}{Open Government Licence - Alberta}
}
\keyword{datasets}
