% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbourecruit.R
\docType{data}
\name{bbourecruit_sim1}
\alias{bbourecruit_sim1}
\title{Simulated Data for Scenario 1}
\format{
A tibble with columns:
\describe{
\item{PopulationName}{Name of the population}
\item{Year}{The year the observation occurred}
\item{Month}{The month the observation occurred}
\item{Day}{The day the observation occurred}
\item{Cows}{The number of cows counted in the group}
\item{Bulls}{The number of bulls counted in the group}
\item{UnknownAdults}{The number of adults which could not be sexed in the
group}
\item{Yearlings}{The number of yearlings counted in the group}
\item{Calves}{The number of calves counted in the group}
}
}
\usage{
bbourecruit_sim1
}
\description{
Data are simulated with \href{https://poissonconsulting.github.io/bbousims/}{\code{bbousims}} package.
This is a stable population spanning 20 years with annual variation on
female adult survival and female calf survival.
Coverage is low, with 20\% of groups observed and 10 collars.
}
\keyword{datasets}
