% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbourecruit.R
\docType{data}
\name{bbourecruit_sim3}
\alias{bbourecruit_sim3}
\title{Simulated Data for Scenario 3}
\format{
A tibble with columns:
\describe{
\item{PopulationName}{Name of the population}
\item{Year}{The year the observation occurred}
\item{Month}{The month the observation occurred}
\item{Day}{The day the observation occurred}
\item{Cows}{The number of cows counted in the group}
\item{Bulls}{The number of bulls counted in the group}
\item{UnknownAdults}{The number of adults which could not be sexed in the
group}
\item{Yearlings}{The number of yearlings counted in the group}
\item{Calves}{The number of calves counted in the group}
}
}
\usage{
bbourecruit_sim3
}
\description{
Data are simulated with \href{https://poissonconsulting.github.io/bbousims/}{\code{bbousims}} package.
This is a population spanning 20 years with negative trend on female adult survival
and additional annual variation on female adult survival and female calf survival.
Coverage is medium, with 50\% of groups observed and 25 collars.
}
\keyword{datasets}
