% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbousurv.R
\docType{data}
\name{bbousurv}
\alias{bbousurv}
\title{Sample Boreal Caribou Survival Data}
\format{
A tibble
\describe{
\item{PopulationName}{Name of the population}
\item{Year}{The year}
\item{Month}{The month}
\item{StartTotal}{The number of living caribou at the start of the month}
\item{MortalitiesCertain}{The number of confirmed caribou mortalities in the
month}
\item{MortalitiesUncertain}{The number of unconfirmed caribou mortalities in
the month}
}
}
\usage{
bbousurv
}
\description{
A data set containing survival information for Boreal Caribou.
}
\keyword{datasets}
