% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbousurv.R
\docType{data}
\name{bbousurv_a}
\alias{bbousurv_a}
\title{Sample Data for Population A}
\format{
A tibble with columns:
\describe{
\item{PopulationName}{Name of the population}
\item{Year}{The year the observation occurred}
\item{Month}{The month the observation occurred}
\item{StartTotal}{The total number of collared caribou at the start of the
month}
\item{MortalitiesCertain}{The number of confirmed caribou mortalities in the
month}
\item{MortalitiesUncertain}{The total number of mortalities that were not
confirmed in that month}
}
}
\usage{
bbousurv_a
}
\description{
The data contains anonymized survival information for boreal caribou population 'A',
which spans 31 years. This dataset follows the methodology described in DeCesare et al. (2012) for estimating
population growth rates using adult female survival data.
}
\details{
The data is released under the \href{https://open.alberta.ca/licence}{Open Government Licence - Alberta}
}
\references{
DeCesare, N.J., Hebblewhite, M., Bradley, M., Smith, K.G., Hervieux, D., & Neufeld, L. (2012).
Estimating ungulate recruitment and growth rates using age ratios.
The Journal of Wildlife Management, 76(1), 144-153. doi:10.1002/jwmg.244
}
\keyword{datasets}
