% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbousurv.R
\docType{data}
\name{bbousurv_sim2}
\alias{bbousurv_sim2}
\title{Simulated Data for Scenario 2}
\format{
A tibble with columns:
\describe{
\item{PopulationName}{Name of the population}
\item{Year}{The year the observation occurred}
\item{Month}{The month the observation occurred}
\item{StartTotal}{The total number of collared caribou at the start of the
month}
\item{MortalitiesCertain}{The number of confirmed caribou mortalities in the
month}
\item{MortalitiesUncertain}{The total number of mortalities that were not
confirmed in that month}
}
}
\usage{
bbousurv_sim2
}
\description{
Data are simulated with \code{bbousims} package.
This is a stable population spanning 20 years with annual variation on
female adult survival and female calf survival.
Coverage is high, with 70\% of groups observed and 40 collars.
}
\keyword{datasets}
