% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbousurv.R
\docType{data}
\name{bbousurv_sim4}
\alias{bbousurv_sim4}
\title{Simulated Data for Scenario 4}
\format{
A tibble with columns:
\describe{
\item{PopulationName}{Name of the population}
\item{Year}{The year the observation occurred}
\item{Month}{The month the observation occurred}
\item{StartTotal}{The total number of collared caribou at the start of the
month}
\item{MortalitiesCertain}{The number of confirmed caribou mortalities in the
month}
\item{MortalitiesUncertain}{The total number of mortalities that were not
confirmed in that month}
}
}
\usage{
bbousurv_sim4
}
\description{
Data are simulated with \code{bbousims} package.
This is a stable population spanning 20 years with small population size
(50 initial adult females) and annual variation on female adult survival and female calf survival.
Coverage is medium, with 40\% of groups observed and 20 collars.
}
\keyword{datasets}
