% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-data-recruitment.R
\name{bbd_chk_data_recruitment}
\alias{bbd_chk_data_recruitment}
\title{Check recruitment data structure}
\usage{
bbd_chk_data_recruitment(data, x_name = deparse(substitute(data)))
}
\arguments{
\item{data}{A data frame.}

\item{x_name}{Name of data frame.}
}
\value{
An invisible data frame.
}
\description{
Check recruitment data structure
}
\examples{
bbd_chk_data_recruitment(bbourecruit_a)
}
