% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk-data-survival.R
\name{bbd_chk_data_survival}
\alias{bbd_chk_data_survival}
\title{Check survival data structure}
\format{
The data must follow the requirements:
\describe{
\item{PopulationName}{Name of the herd or population}
\item{Year}{The calendar year the observation occurred. Must be a positive
integer.}
\item{Month}{The calendar month the observation occurred. Must be an integer
between 1 and 12.}
\item{StartTotal}{The total number of collared caribou at the start of the
month. Must be a positive integer.}
\item{MortalitiesCertain}{The total number of confirmed mortalities in that
month. Must be a positive integer.}
\item{MortalitiesUncertain}{The total number of mortalities that were not
confirmed in that month. Must be a positive integer.}
}
}
\usage{
bbd_chk_data_survival(data, x_name = deparse(substitute(data)))
}
\arguments{
\item{data}{A data frame.}

\item{x_name}{Name of data frame.}
}
\value{
An invisible data frame.
}
\description{
The data must follow all requirements to not error. This format is required
for usage of the bbou suite of tools.
}
\examples{
bbd_chk_data_survival(bbousurv_a)
bbd_chk_data_survival(bbousurv_b)
bbd_chk_data_survival(bbousurv_c)

# this example will error as it doesn't follow the requirements
x <- bbousurv_c
x[1, 3] <- 14L
try(bbd_chk_data_survival(x))
}
