% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbousurv.R
\docType{data}
\name{bbousurv_a}
\alias{bbousurv_a}
\title{Sample Data for Population A}
\format{
A tibble
\describe{
\item{PopulationName}{Name of the population}
\item{Year}{The year}
\item{Month}{The month}
\item{StartTotal}{The number of living caribou at the start of the month}
\item{MortalitiesCertain}{The number of confirmed caribou mortalities in the
month}
\item{MortalitiesUncertain}{The number of unconfirmed caribou mortalities in
the month}
}
}
\usage{
bbousurv_a
}
\description{
The data contains survival information for a boreal caribou population A
which spans 30 years.
}
\details{
The data is released under the \href{https://open.alberta.ca/licence}{Open Government Licence - Alberta}
}
\keyword{datasets}
